containers/fbworker/requirements.txt
python-dateutil
pika
rethinkdb


